src/requirements.txt
pystray
pillow
SpeechRecognition
pyaudio
rapidfuzz